requirements.txt
flask==3.1.0
gspread==6.1.4
google-auth==2.38.0
requests==2.32.3
python-dotenv==1.0.1
